requirement.txt
streamlit
opencv-python
numpy
pillow
svgwrite
scikit-image




